requirements.txt
flask
python-dotenv
openai
aiohttp
requests
gunicorn